// src/app/enums/advanced/storage-value.enum.ts
export enum StorageValueEnum {
  USER_DETAILS = 'userDetails',
  LOGIN_TOKEN = 'loginToken',
  PUSH_TOKEN = 'pushToken',
  PAYMENT_DATA = 'paymentData',
  CREDIT_CARD_ENCRYPTED = 'creditCardEncrypted',
  CREDIT_CARD_FOR_DIGITS = 'creditCard4Digits',
  CODE_SENT_DATE = 'codeSentDate',
  PHONE = 'phone',
  EXPIRES_AT = 'expiresAt'
}